+++ microservicios-compras/src/main/java/com/prpymes/microservicios/app/compras/models/repository/ManufacturaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import com.prpymes.microservicios.app.compras.models.entity.Manufactura;
 
-public interface ManufacturaRepository extends PagingAndSortingRepository<Manufactura, Integer> {
+public interface ManufacturaRepository extends PagingAndSortingRepository<Manufactura, Long> {
     
 }
