+++ microservicios-compras/src/main/java/com/prpymes/microservicios/app/compras/services/ProveedorService.java
@@ -8,6 +8,6 @@ import com.prpymes.microservicios.commons.services.CommonServiceC;
 public interface ProveedorService extends CommonServiceC<Proveedor>{
     public List<Proveedor> findByNombre(String term);
 	
-    public List<Object> findPresupuestosForProvedor(Integer idProveedor);
+    public List<Object> findPresupuestosForProvedor(Long idProveedor);
 
 }
